+++ api/datadogV1/model_cancel_downtimes_by_scope_request.go
@@ -80,21 +80,10 @@ func (o CancelDowntimesByScopeRequest) MarshalJSON() ([]byte, error) {
 // UnmarshalJSON deserializes the given payload.
 func (o *CancelDowntimesByScopeRequest) UnmarshalJSON(bytes []byte) (err error) {
 	raw := map[string]interface{}{}
-	required := struct {
+	all := struct {
 		Scope *string `json:"scope"`
 	}{}
-	all := struct {
-		Scope string `json:"scope"`
-	}{}
-	err = json.Unmarshal(bytes, &required)
-	if err != nil {
-		return err
-	}
-	if required.Scope == nil {
-		return fmt.Errorf("required field scope missing")
-	}
-	err = json.Unmarshal(bytes, &all)
-	if err != nil {
+	if err = json.Unmarshal(bytes, &all); err != nil {
 		err = json.Unmarshal(bytes, &raw)
 		if err != nil {
 			return err
@@ -102,13 +91,16 @@ func (o *CancelDowntimesByScopeRequest) UnmarshalJSON(bytes []byte) (err error) 
 		o.UnparsedObject = raw
 		return nil
 	}
+	if all.Scope == nil {
+		return fmt.Errorf("required field scope missing")
+	}
 	additionalProperties := make(map[string]interface{})
 	if err = json.Unmarshal(bytes, &additionalProperties); err == nil {
 		datadog.DeleteKeys(additionalProperties, &[]string{"scope"})
 	} else {
 		return err
 	}
-	o.Scope = all.Scope
+	o.Scope = *all.Scope
 	if len(additionalProperties) > 0 {
 		o.AdditionalProperties = additionalProperties
 	}
