+++ api/datadogV1/model_event_query_definition.go
@@ -105,26 +105,11 @@ func (o EventQueryDefinition) MarshalJSON() ([]byte, error) {
 // UnmarshalJSON deserializes the given payload.
 func (o *EventQueryDefinition) UnmarshalJSON(bytes []byte) (err error) {
 	raw := map[string]interface{}{}
-	required := struct {
+	all := struct {
 		Search        *string `json:"search"`
 		TagsExecution *string `json:"tags_execution"`
 	}{}
-	all := struct {
-		Search        string `json:"search"`
-		TagsExecution string `json:"tags_execution"`
-	}{}
-	err = json.Unmarshal(bytes, &required)
-	if err != nil {
-		return err
-	}
-	if required.Search == nil {
-		return fmt.Errorf("required field search missing")
-	}
-	if required.TagsExecution == nil {
-		return fmt.Errorf("required field tags_execution missing")
-	}
-	err = json.Unmarshal(bytes, &all)
-	if err != nil {
+	if err = json.Unmarshal(bytes, &all); err != nil {
 		err = json.Unmarshal(bytes, &raw)
 		if err != nil {
 			return err
@@ -132,14 +117,20 @@ func (o *EventQueryDefinition) UnmarshalJSON(bytes []byte) (err error) {
 		o.UnparsedObject = raw
 		return nil
 	}
+	if all.Search == nil {
+		return fmt.Errorf("required field search missing")
+	}
+	if all.TagsExecution == nil {
+		return fmt.Errorf("required field tags_execution missing")
+	}
 	additionalProperties := make(map[string]interface{})
 	if err = json.Unmarshal(bytes, &additionalProperties); err == nil {
 		datadog.DeleteKeys(additionalProperties, &[]string{"search", "tags_execution"})
 	} else {
 		return err
 	}
-	o.Search = all.Search
-	o.TagsExecution = all.TagsExecution
+	o.Search = *all.Search
+	o.TagsExecution = *all.TagsExecution
 	if len(additionalProperties) > 0 {
 		o.AdditionalProperties = additionalProperties
 	}
