+++ api/datadogV1/model_list_stream_group_by_items.go
@@ -78,21 +78,10 @@ func (o ListStreamGroupByItems) MarshalJSON() ([]byte, error) {
 // UnmarshalJSON deserializes the given payload.
 func (o *ListStreamGroupByItems) UnmarshalJSON(bytes []byte) (err error) {
 	raw := map[string]interface{}{}
-	required := struct {
+	all := struct {
 		Facet *string `json:"facet"`
 	}{}
-	all := struct {
-		Facet string `json:"facet"`
-	}{}
-	err = json.Unmarshal(bytes, &required)
-	if err != nil {
-		return err
-	}
-	if required.Facet == nil {
-		return fmt.Errorf("required field facet missing")
-	}
-	err = json.Unmarshal(bytes, &all)
-	if err != nil {
+	if err = json.Unmarshal(bytes, &all); err != nil {
 		err = json.Unmarshal(bytes, &raw)
 		if err != nil {
 			return err
@@ -100,13 +89,16 @@ func (o *ListStreamGroupByItems) UnmarshalJSON(bytes []byte) (err error) {
 		o.UnparsedObject = raw
 		return nil
 	}
+	if all.Facet == nil {
+		return fmt.Errorf("required field facet missing")
+	}
 	additionalProperties := make(map[string]interface{})
 	if err = json.Unmarshal(bytes, &additionalProperties); err == nil {
 		datadog.DeleteKeys(additionalProperties, &[]string{"facet"})
 	} else {
 		return err
 	}
-	o.Facet = all.Facet
+	o.Facet = *all.Facet
 	if len(additionalProperties) > 0 {
 		o.AdditionalProperties = additionalProperties
 	}
