+++ api/datadogV1/model_notebook_absolute_time.go
@@ -147,27 +147,12 @@ func (o NotebookAbsoluteTime) MarshalJSON() ([]byte, error) {
 // UnmarshalJSON deserializes the given payload.
 func (o *NotebookAbsoluteTime) UnmarshalJSON(bytes []byte) (err error) {
 	raw := map[string]interface{}{}
-	required := struct {
+	all := struct {
 		End   *time.Time `json:"end"`
+		Live  *bool      `json:"live,omitempty"`
 		Start *time.Time `json:"start"`
 	}{}
-	all := struct {
-		End   time.Time `json:"end"`
-		Live  *bool     `json:"live,omitempty"`
-		Start time.Time `json:"start"`
-	}{}
-	err = json.Unmarshal(bytes, &required)
-	if err != nil {
-		return err
-	}
-	if required.End == nil {
-		return fmt.Errorf("required field end missing")
-	}
-	if required.Start == nil {
-		return fmt.Errorf("required field start missing")
-	}
-	err = json.Unmarshal(bytes, &all)
-	if err != nil {
+	if err = json.Unmarshal(bytes, &all); err != nil {
 		err = json.Unmarshal(bytes, &raw)
 		if err != nil {
 			return err
@@ -175,15 +160,21 @@ func (o *NotebookAbsoluteTime) UnmarshalJSON(bytes []byte) (err error) {
 		o.UnparsedObject = raw
 		return nil
 	}
+	if all.End == nil {
+		return fmt.Errorf("required field end missing")
+	}
+	if all.Start == nil {
+		return fmt.Errorf("required field start missing")
+	}
 	additionalProperties := make(map[string]interface{})
 	if err = json.Unmarshal(bytes, &additionalProperties); err == nil {
 		datadog.DeleteKeys(additionalProperties, &[]string{"end", "live", "start"})
 	} else {
 		return err
 	}
-	o.End = all.End
+	o.End = *all.End
 	o.Live = all.Live
-	o.Start = all.Start
+	o.Start = *all.Start
 	if len(additionalProperties) > 0 {
 		o.AdditionalProperties = additionalProperties
 	}
