+++ api/datadogV1/model_pager_duty_service_key.go
@@ -78,21 +78,10 @@ func (o PagerDutyServiceKey) MarshalJSON() ([]byte, error) {
 // UnmarshalJSON deserializes the given payload.
 func (o *PagerDutyServiceKey) UnmarshalJSON(bytes []byte) (err error) {
 	raw := map[string]interface{}{}
-	required := struct {
+	all := struct {
 		ServiceKey *string `json:"service_key"`
 	}{}
-	all := struct {
-		ServiceKey string `json:"service_key"`
-	}{}
-	err = json.Unmarshal(bytes, &required)
-	if err != nil {
-		return err
-	}
-	if required.ServiceKey == nil {
-		return fmt.Errorf("required field service_key missing")
-	}
-	err = json.Unmarshal(bytes, &all)
-	if err != nil {
+	if err = json.Unmarshal(bytes, &all); err != nil {
 		err = json.Unmarshal(bytes, &raw)
 		if err != nil {
 			return err
@@ -100,13 +89,16 @@ func (o *PagerDutyServiceKey) UnmarshalJSON(bytes []byte) (err error) {
 		o.UnparsedObject = raw
 		return nil
 	}
+	if all.ServiceKey == nil {
+		return fmt.Errorf("required field service_key missing")
+	}
 	additionalProperties := make(map[string]interface{})
 	if err = json.Unmarshal(bytes, &additionalProperties); err == nil {
 		datadog.DeleteKeys(additionalProperties, &[]string{"service_key"})
 	} else {
 		return err
 	}
-	o.ServiceKey = all.ServiceKey
+	o.ServiceKey = *all.ServiceKey
 	if len(additionalProperties) > 0 {
 		o.AdditionalProperties = additionalProperties
 	}
