+++ api/datadogV1/model_run_workflow_widget_input.go
@@ -105,26 +105,11 @@ func (o RunWorkflowWidgetInput) MarshalJSON() ([]byte, error) {
 // UnmarshalJSON deserializes the given payload.
 func (o *RunWorkflowWidgetInput) UnmarshalJSON(bytes []byte) (err error) {
 	raw := map[string]interface{}{}
-	required := struct {
+	all := struct {
 		Name  *string `json:"name"`
 		Value *string `json:"value"`
 	}{}
-	all := struct {
-		Name  string `json:"name"`
-		Value string `json:"value"`
-	}{}
-	err = json.Unmarshal(bytes, &required)
-	if err != nil {
-		return err
-	}
-	if required.Name == nil {
-		return fmt.Errorf("required field name missing")
-	}
-	if required.Value == nil {
-		return fmt.Errorf("required field value missing")
-	}
-	err = json.Unmarshal(bytes, &all)
-	if err != nil {
+	if err = json.Unmarshal(bytes, &all); err != nil {
 		err = json.Unmarshal(bytes, &raw)
 		if err != nil {
 			return err
@@ -132,14 +117,20 @@ func (o *RunWorkflowWidgetInput) UnmarshalJSON(bytes []byte) (err error) {
 		o.UnparsedObject = raw
 		return nil
 	}
+	if all.Name == nil {
+		return fmt.Errorf("required field name missing")
+	}
+	if all.Value == nil {
+		return fmt.Errorf("required field value missing")
+	}
 	additionalProperties := make(map[string]interface{})
 	if err = json.Unmarshal(bytes, &additionalProperties); err == nil {
 		datadog.DeleteKeys(additionalProperties, &[]string{"name", "value"})
 	} else {
 		return err
 	}
-	o.Name = all.Name
-	o.Value = all.Value
+	o.Name = *all.Name
+	o.Value = *all.Value
 	if len(additionalProperties) > 0 {
 		o.AdditionalProperties = additionalProperties
 	}
